Added edit inventory name functionality, moved delete item button to item form
